Use long instead of int for key type

+++ GrainAuger.Benchmarks/ProcessChains/Jobs.cs
@@ -7,7 +7,7 @@ public class Jobs
     [AugerJobConfiguration("SplitJob")]
     public void ConfigureSplit(IAugerJobBuilder builder)
     {
-        var input1 = builder.FromStream<int, int>("Memory", "InputSplit");
+        var input1 = builder.FromStream<int, long>("Memory", "InputSplit");
         var split1 = input1.Process<PassProcessor>();
         var split2 = split1.Process<PassProcessor>();
     }
@@ -15,7 +15,7 @@ public class Jobs
     [AugerJobConfiguration("JoinedJob")]
     public void ConfigureJoined(IAugerJobBuilder builder)
     {
-        var input2 = builder.FromStream<int, int>("Memory", "InputJoin");
+        var input2 = builder.FromStream<int, long>("Memory", "InputJoin");
         var joined = input2.Process<PassProcessor, PassProcessor>();
     }
 }
